MapExplorer moved to its own folder
